Update README.md
I have created a complete Amazon clone, which is a copy of the popular online
shopping platform., meticulously mirroring the functionalities of the renowned e-commerce giant. The project encapsulates a dynamic product catalog, an intuitive
search bar, and a seamless shopping cart. Developed with precision using React for a
reactive front end, Firebase for robust backend support, and Tailwind CSS for a polished user interface.